Move code from controller to module
